refactor: improve code organization and Arc usage patterns
Major refactorings:
- Move CommandRegistry to top-level commands module (lua_registry.rs)
  * Better separation of concerns - registry is not Lua execution-specific
  * Makes it accessible to both Lua and native commands

- Add load_lua_file() helper in execute/mod.rs
  * Eliminates duplicated error handling code
  * Makes script loading more maintainable

- Fix Arc usage patterns throughout codebase
  * Change build_handlers to accept Arc<T> instead of &Arc<T>
  * More idiomatic Rust - Arc is cheap to clone, no need to borrow it
  * Clarifies ownership semantics

- Update all imports to use new CommandRegistry location:
  * main.rs
  * reload.rs
  * lua_command.rs
  * execute/extensions/discord.rs

All changes compile cleanly and pass clippy with -D warnings.
