+++ src/commands/execute/mod.rs
@@ -147,6 +147,16 @@ impl Handler {
     }
 }
 
+fn load_lua_file(lua: &mlua::Lua, path: &str) -> mlua::Result<()> {
+    let script_path = std::path::Path::new(path);
+    if script_path.exists() {
+        let code = std::fs::read_to_string(script_path)
+            .map_err(|e| mlua::Error::RuntimeError(format!("Failed to read {}: {}", path, e)))?;
+        lua.load(&code).set_name(path).exec()?;
+    }
+    Ok(())
+}
+
 fn create_lua_state(
     ai: Arc<Ai>,
     currency_converter: Arc<CurrencyConverter>,
@@ -162,14 +172,7 @@ fn create_lua_state(
     )?;
 
     extensions::register(&lua, ai, currency_converter, output_tx, print_tx)?;
-
-    // Load scripts/main.lua if it exists
-    let main_script = std::path::Path::new("scripts/main.lua");
-    if main_script.exists() {
-        let code = std::fs::read_to_string(main_script)
-            .map_err(|e| mlua::Error::RuntimeError(format!("Failed to read main.lua: {}", e)))?;
-        lua.load(&code).set_name("main.lua").exec()?;
-    }
+    load_lua_file(&lua, "scripts/main.lua")?;
 
     Ok(lua)
 }
@@ -179,7 +182,7 @@ pub fn create_global_lua_state(
     currency_converter: Arc<CurrencyConverter>,
     output_tx: flume::Sender<String>,
     print_tx: flume::Sender<String>,
-    command_registry: extensions::discord::CommandRegistry,
+    command_registry: crate::commands::CommandRegistry,
 ) -> mlua::Result<mlua::Lua> {
     let lua = mlua::Lua::new_with(
         {
@@ -192,22 +195,8 @@ pub fn create_global_lua_state(
     extensions::register(&lua, ai, currency_converter, output_tx, print_tx)?;
     extensions::discord::register(&lua, command_registry)?;
 
-    // Load scripts/main.lua if it exists
-    let main_script = std::path::Path::new("scripts/main.lua");
-    if main_script.exists() {
-        let code = std::fs::read_to_string(main_script)
-            .map_err(|e| mlua::Error::RuntimeError(format!("Failed to read main.lua: {}", e)))?;
-        lua.load(&code).set_name("main.lua").exec()?;
-    }
-
-    // Load scripts/commands.lua if it exists
-    let commands_script = std::path::Path::new("scripts/commands.lua");
-    if commands_script.exists() {
-        let code = std::fs::read_to_string(commands_script).map_err(|e| {
-            mlua::Error::RuntimeError(format!("Failed to read commands.lua: {}", e))
-        })?;
-        lua.load(&code).set_name("commands.lua").exec()?;
-    }
+    load_lua_file(&lua, "scripts/main.lua")?;
+    load_lua_file(&lua, "scripts/commands.lua")?;
 
     Ok(lua)
 }
