+++ src/commands/mod.rs
@@ -2,7 +2,10 @@ use serenity::all::{CommandInteraction, Http};
 
 pub mod execute;
 pub mod lua_command;
+pub mod lua_registry;
 pub mod reload;
+
+pub use lua_registry::CommandRegistry;
 
 #[serenity::async_trait]
 pub trait CommandHandler: Send + Sync {
